Refactoring modules at a better places
