+++ pyFAI/gui/calibration/CalibrationContext.py
@@ -25,7 +25,7 @@
 
 __authors__ = ["V. Valls"]
 __license__ = "MIT"
-__date__ = "21/08/2018"
+__date__ = "24/08/2018"
 
 import weakref
 import logging
@@ -40,7 +40,7 @@ from .model.CalibrationModel import CalibrationModel
 from .model.DataModel import DataModel
 from . import utils
 from ..utils import eventutils
-from . import units
+from ..utils import units
 
 
 _logger = logging.getLogger(__name__)
