+++ pyFAI/gui/calibration/IntegrationTask.py
@@ -27,7 +27,7 @@ from __future__ import absolute_import
 
 __authors__ = ["V. Valls"]
 __license__ = "MIT"
-__date__ = "23/08/2018"
+__date__ = "24/08/2018"
 
 import logging
 import numpy
@@ -45,7 +45,7 @@ from . import utils
 from .model.DataModel import DataModel
 from ..widgets.QuantityLabel import QuantityLabel
 from .CalibrationContext import CalibrationContext
-from . import units
+from ..utils import units
 from .helper.MarkerManager import MarkerManager
 
 _logger = logging.getLogger(__name__)
