+++ pyFAI/gui/utils/units.py
