+++ pyFAI/gui/utils/validators.py
