+++ pyFAI/gui/widgets/CalibrantPreview.py
