+++ pyFAI/gui/widgets/DetectorLabel.py
