+++ pyFAI/gui/widgets/DetectorModel.py
