+++ pyFAI/gui/widgets/FileEdit.py
