+++ pyFAI/gui/widgets/UnitLabel.py
@@ -25,13 +25,13 @@
 
 __copyright__ = "European Synchrotron Radiation Facility, Grenoble, France"
 __license__ = "MIT"
-__date__ = "22/08/2018"
+__date__ = "24/08/2018"
 
 import functools
 
 from silx.gui import qt
 
-from ..calibration import units
+from ..utils import units
 
 
 class UnitLabel(qt.QLabel):
