+++ pyFAI/gui/widgets/UnitSelector.py
@@ -27,37 +27,57 @@ from __future__ import absolute_import
 
 __authors__ = ["V. Valls"]
 __license__ = "MIT"
-__date__ = "07/08/2018"
+__date__ = "24/08/2018"
 
 from silx.gui import qt
-import pyFAI.detectors
-from .model.DetectorModel import DetectorModel
+from ..calibration.model.DataModel import DataModel
 
 
-class DetectorManufacturer(qt.QComboBox):
+class UnitSelector(qt.QComboBox):
 
     def __init__(self, parent=None):
-        super(DetectorManufacturer, self).__init__(parent)
-
-        # feed the widget with default manufacturers
-        manufacturers = set([])
-        for detector in pyFAI.detectors.ALL_DETECTORS.values():
-            manufacturers.add(detector.MANUFACTURER)
-
-        hasOther = None in manufacturers
-        manufacturers.remove(None)
-        manufacturers = sorted(list(manufacturers))
-
-        for manufacturer in manufacturers:
-            self.addItem(manufacturer, manufacturer)
-        if hasOther:
-            self.addItem("Others", None)
-        self.insertItem(0, "Any", "*")
-        self.insertSeparator(1)
+        super(UnitSelector, self).__init__(parent)
 
         self.__model = None
-        self.setModel(DetectorModel())
-        self.setCurrentIndex(0)
+        self.setModel(DataModel())
+        self.currentIndexChanged[int].connect(self.__currentIndexChanged)
+
+    def setUnits(self, units):
+        previousUnit = self.__model.value()
+        old = self.blockSignals(True)
+        # clean up
+        self.clear()
+
+        units = sorted(list(units), key=lambda u: u.name)
+
+        for unit in units:
+            self.addItem(unit.name, unit)
+        # try to find the previous unit in the new list
+        if previousUnit is None:
+            currentIndex = self.currentIndex()
+            index = -1
+        else:
+            currentIndex = self.currentIndex()
+            index = self.findUnit(previousUnit)
+        if index == -1:
+            self.blockSignals(old)
+            if currentIndex != index:
+                # the previous selected unit is not anymore available
+                self.setCurrentIndex(index)
+        else:
+            # the previous index is found
+            # we dont have to emit signals
+            self.setCurrentIndex(index)
+            self.blockSignals(old)
+
+    def __currentIndexChanged(self, index):
+        model = self.model()
+        if model is None:
+            return
+        unit = self.itemData(index)
+        old = self.blockSignals(True)
+        model.setValue(unit)
+        self.blockSignals(old)
 
     def setModel(self, model):
         if self.__model is not None:
@@ -67,37 +87,26 @@ class DetectorManufacturer(qt.QComboBox):
             self.__model.changed.connect(self.__modelChanged)
         self.__modelChanged()
 
-    def currentManufacturer(self):
-        index = self.currentIndex()
-        if index == -1:
-            return "*"
-        manufacturer = self.itemData(index)
-        return manufacturer
-
-    def findManufacturer(self, manufacturer):
-        """Returns the first index containing the requested manufacturer.
+    def findUnit(self, unit):
+        """Returns the first index containing the requested detector.
         Else return -1"""
         for index in range(self.count()):
             item = self.itemData(index)
-            if item == "*":
-                continue
-            if item is manufacturer:
+            if item is unit:
                 return index
-        # TODO we are supposed to return other group
-        # or create a new manufacturer
         return -1
 
     def __modelChanged(self):
-        value = self.__model.detector()
+        value = self.__model.value()
         if value is None:
             self.setCurrentIndex(-1)
         else:
-            manufacturer = value.MANUFACTURER
+            unit = value
             index = self.currentIndex()
             item = self.itemData(index)
-            if item != manufacturer:
+            if item is not unit:
                 # findData is not working
-                index = self.findManufacturer(manufacturer)
+                index = self.findUnit(unit)
                 self.setCurrentIndex(index)
 
     def model(self):
